requirements.txt
streamlit
langchain
python-dotenv
pypdf2
huggingface_hub
faiss-cpu
InstructorEmbedding
sentence_transformers
requests
faiss-cpu
streamlit_chat